feat: 增加移动端重构接口

+++ sync.js
@@ -1,15 +1,18 @@
-var sequelize;
+var sequelize = require('./src/db')[0];
 
-sequelize = require('./src/db')[0];
+var argv = process.argv || [],
+  command = argv[2] || '',
+  isForce = command.indexOf('force') !== -1;
 
-console.log('Drop all schemas.');
-
-sequelize.drop();
+if (isForce) {
+  console.log('Drop all schemas.');
+  sequelize.drop();
+}
 
 console.log('Sync all schemas.');
 
 sequelize.sync({
-  force: true
+  force: isForce
 }).then(function() {
   return console.log('All done!');
 })["catch"](function(err) {
